Added latest test example

+++ 2025-05-14-unitTesting/factorial.h
@@ -1,2 +1,3 @@
 #pragma once
-int factorial(int n);
+#include <iostream>
+long factorial(long n);
